requirements.txt
PyYAML==6.0.3
requests==2.32.5

